requirements.txt
scikit-learn
numpy
pandas
pymongo
requests
flask
python-dotenv
